# improve performance of ValueRetrievers