--- a/TechTalk.SpecFlow/Assist/ValueRetrievers/CharValueRetriever.cs
+++ b/TechTalk.SpecFlow/Assist/ValueRetrievers/CharValueRetriever.cs
@@ -1,18 +1,12 @@
 ﻿namespace TechTalk.SpecFlow.Assist.ValueRetrievers
 {
-    public class CharValueRetriever : NonNullableValueRetriever<char>
+    public class CharValueRetriever : StructRetriever<char>
     {
-        public override char GetValue(string value)
+        protected override char GetNonEmptyValue(string value)
         {
-            return ThisStringIsNotASingleCharacter(value)
-                       ? '\0'
-                       : value[0];
-        }
-
-
-        private bool ThisStringIsNotASingleCharacter(string value)
-        {
-            return string.IsNullOrEmpty(value) || value.Length > 1;
+            return value != null && value.Length == 1 
+                ? value[0] 
+                : '\0';
         }
     }
 }