--- a/TechTalk.SpecFlow/Assist/ValueRetrievers/DateTimeValueRetriever.cs
+++ b/TechTalk.SpecFlow/Assist/ValueRetrievers/DateTimeValueRetriever.cs
@@ -3,9 +3,9 @@ using System.Globalization;
 
 namespace TechTalk.SpecFlow.Assist.ValueRetrievers
 {
-    public class DateTimeValueRetriever : NonNullableValueRetriever<DateTime>
+    public class DateTimeValueRetriever : StructRetriever<DateTime>
     {
-        public override DateTime GetValue(string value)
+        protected override DateTime GetNonEmptyValue(string value)
         {
             DateTime.TryParse(value, CultureInfo.CurrentCulture, DateTimeStyles.None, out DateTime returnValue);
             return returnValue;