--- a/TechTalk.SpecFlow/Assist/ValueRetrievers/DoubleValueRetriever.cs
+++ b/TechTalk.SpecFlow/Assist/ValueRetrievers/DoubleValueRetriever.cs
@@ -3,9 +3,9 @@ using System.Globalization;
 
 namespace TechTalk.SpecFlow.Assist.ValueRetrievers
 {
-    public class DoubleValueRetriever : NonNullableValueRetriever<double>
+    public class DoubleValueRetriever : StructRetriever<double>
     {
-        public override double GetValue(string value)
+        protected override double GetNonEmptyValue(string value)
         {
             Double.TryParse(value, NumberStyles.Any, CultureInfo.CurrentCulture, out double returnValue);
             return returnValue;