… TechTalk.SpecFlow/Assist/ValueRetrievers/NonNullableValueRetriever.cs
… TechTalk.SpecFlow/Assist/ValueRetrievers/NullValueRetriever.cs
… TechTalk.SpecFlow/Assist/ValueRetrievers/NullableBoolValueRetriever.cs
… TechTalk.SpecFlow/Assist/ValueRetrievers/NullableByteValueRetriever.cs
… TechTalk.SpecFlow/Assist/ValueRetrievers/NullableCharValueRetriever.cs
… TechTalk.SpecFlow/Assist/ValueRetrievers/NullableDateTimeOffsetValueRetriever.cs
… TechTalk.SpecFlow/Assist/ValueRetrievers/NullableDateTimeValueRetriever.cs
… TechTalk.SpecFlow/Assist/ValueRetrievers/NullableDecimalValueRetriever.cs
… TechTalk.SpecFlow/Assist/ValueRetrievers/NullableDoubleValueRetriever.cs
… TechTalk.SpecFlow/Assist/ValueRetrievers/NullableFloatValueRetriever.cs
… TechTalk.SpecFlow/Assist/ValueRetrievers/NullableGuidValueRetriever.cs
… TechTalk.SpecFlow/Assist/ValueRetrievers/NullableIntValueRetriever.cs
… TechTalk.SpecFlow/Assist/ValueRetrievers/NullableLongValueRetriever.cs
… TechTalk.SpecFlow/Assist/ValueRetrievers/NullableSByteValueRetriever.cs
… TechTalk.SpecFlow/Assist/ValueRetrievers/NullableShortValueRetriever.cs
… TechTalk.SpecFlow/Assist/ValueRetrievers/NullableTimeSpanValueRetriever.cs
… TechTalk.SpecFlow/Assist/ValueRetrievers/NullableUIntValueRetriever.cs
… TechTalk.SpecFlow/Assist/ValueRetrievers/NullableULongValueRetriever.cs
… TechTalk.SpecFlow/Assist/ValueRetrievers/NullableUShortValueRetriever.cs
… TechTalk.SpecFlow/Assist/ValueRetrievers/NullableValueRetriever.cs
… TechTalk.SpecFlow/Assist/ValueRetrievers/SByteValueRetriever.cs
… TechTalk.SpecFlow/Assist/ValueRetrievers/ShortValueRetriever.cs
… TechTalk.SpecFlow/Assist/ValueRetrievers/StringValueRetriever.cs
… TechTalk.SpecFlow/Assist/ValueRetrievers/StructRetriever.cs
… TechTalk.SpecFlow/Assist/ValueRetrievers/TimeSpanValueRetriever.cs
… TechTalk.SpecFlow/Assist/ValueRetrievers/UIntValueRetriever.cs
… TechTalk.SpecFlow/Assist/ValueRetrievers/ULongValueRetriever.cs
… TechTalk.SpecFlow/Assist/ValueRetrievers/UShortValueRetriever.cs
… TechTalk.SpecFlow/Assist/ValueRetrievers/UriValueRetriever.cs
… Tests/TechTalk.SpecFlow.RuntimeTests/AssistTests/ServiceTests.cs