Hey

In my project I am using the TableHelperExtensionMethods quite extensively and my analysis showed that they're not the fastest, so I wanted to have a look whether I can improve things a bit.
I started by improving the performance of the ValueRetrievers to get a feeling and to see if this is something you want to include.

Regarding the changes in this PR.

- I've tried to keep the public interface as is. 
- The functionality should be the same as before. 
- I've restructured / simplified  the tests a bit, but kept the test data and extended with a few new ones for missing cases.

Let me know what you think.
Cheers

## Types of changes
- [x] Bug fix (non-breaking change which fixes an issue).
- [ ] New feature (non-breaking change which adds functionality).
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected).

## Checklist:
- [x] I've added tests for my code. (most of the time mandatory)
- [x] I have added an entry to the changelog. (mandatory)
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.