@SabotageAndi 2020-03-30
Thanks for the PR. It's quite big so we will need some time to review it.

It would be interesting how much the performance was improved with these changes. Could you provide some numbers for us?