@bollhals 2020-03-30
> Thanks for the PR. It's quite big so we will need some time to review it.
>
> It would be interesting how much the performance was improved with these changes. Could you provide some numbers for us?
Sure, no problem, just comment if you have any questions or feedback. 

There are two possible ways to get some numbers.
1. Via profiling the existing unit tests.
2. Via a new benchmarking project and usage of BenchmarkDotNet. 

Regarding 1, it's difficult to get proper numbers as the general test execute too quickly (< 2ms). I can give you one example of the following test, as this one takes some time to execute:
ListRetrieverTests.Retrieves_value_enumeration_from_comma_separated_list (scoped to the Retrieve Method, which is the relevant method to test)
Old 172 ms:
![image](https://user-images.githubusercontent.com/2437596/77943004-cf909a80-72bc-11ea-8797-b3b349cd7ce9.png)
New 148 ms:
![image](https://user-images.githubusercontent.com/2437596/77943079-e8994b80-72bc-11ea-8410-86406ffd3f53.png)
~24 ms less, ~14% less time spent

Regarding 2, if you're open for it, I can also set up a new project and throw in some benchmarks so we can see it more closely.