@bollhals 2020-03-30
Maybe this test also highlights some benefits:
CreateInstanceHelperMethodTests.Can_create_an_instance_with_a_constructor:
Old 46 ms:
![image](https://user-images.githubusercontent.com/2437596/77943686-ec799d80-72bd-11ea-86c9-fe34a3a59b17.png)

New 38ms:
![image](https://user-images.githubusercontent.com/2437596/77943634-d1a72900-72bd-11ea-8ae2-ab76275a09f2.png)

~8 ms less, ~18% less time spent

Keep in mind though, profiling tests always means we're only seeing full initialization time (so including JIT, static ctors, etc)
Some of my changes target the init time, so the benefits will be less if the same function gets called multiple times / during a full test run.