@304NotModified 2020-04-26
While I'm not very impressed with the performance improvement, I really like that a lot of code has been deleted (I'm serious). Less code = less maintenance and less bugs. Also I think it's really nice that there are many tests cases added.