@godrose 2020-05-08
Hooray!

On Fri, 8 May 2020 at 15:16 bollhals <notifications@github.com> wrote:

> *@bollhals* commented on this pull request.
> ------------------------------
>
> In changelog.txt
> <https://github.com/SpecFlowOSS/SpecFlow/pull/1928#discussion_r422109939>:
>
> > @@ -2,6 +2,7 @@ Changes since 3.1.89
>
>  Fixes:
>  + Update the default C# skeleton template to use context injection instead of the deprecated ScenarioContext.Current
> ++ Improved performance of ValueRetrievers
>
> Done.
> I think I got all the testcases you've introduced in #1941
> <https://github.com/SpecFlowOSS/SpecFlow/pull/1941> But would you mind
> doublechecking?
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/SpecFlowOSS/SpecFlow/pull/1928#discussion_r422109939>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AB3OIIZFWVKM374NJUF3TC3RQPZ3FANCNFSM4LVYHGLQ>
> .
>
-- 
Best Regards,
Gennady Verdel