@SabotageAndi 2020-05-20
Thanks for fixing the merge conflicts and for your contribution to SpecFlow. 

If you would like us to send you some SpecFlow stickers as a thank you, please fill out the form [here](https://specflow.org/contributer-package/).